# event-manager/src/main/resources/static/event-manager-openapi.yaml
...
openapi: 3.0.0
info:
  description: API for managing event locations
  title: Event Management Service
  version: 1.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
  "/locations":
    get:
      description: |-
        Получение списка всех локаций.

        Изменения по итерациям:
        - Итерация 1: добавить endpoint для получения списка локаций.
        - Итерация 2: endpoint теперь вызывается только с Bearer JWT; доступ разрешен ролям ADMIN и USER.
        - Итерация 5: изменена внутренняя логика чтения; используется cache-aside для списка локаций с Redis.
      parameters:
        - name: pageNum
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 2
            maximum: 20
            minimum: 2
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  "$ref": "#/components/schemas/EventLocation"
                type: array
          description: Список всех локаций
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Получить список всех локаций. Allowed roles=[ADMIN, USER]
      tags:
      - Locations
    post:
      description: |-
        Создать новую локацию. Id должен задаваться на стороне приложения (базой данных).

        Изменения по итерациям:
        - Итерация 1: добавить endpoint для создания локации с валидацией request body и возвратом созданной сущности.
        - Итерация 2: endpoint теперь вызывается только с Bearer JWT; доступ разрешен только роли ADMIN.
        - Итерация 5: изменена внутренняя логика после записи; добавлена инвалидация кешей локаций.
      requestBody:
        content:
          application/json:
            example:
              name: Ледовый дворец
              address: г. СПб, Пятилеток 1
              capacity: 12300
              description: Спортивно-концертный комплекс в Санкт-Петербурге
            schema:
              "$ref": "#/components/schemas/EventLocation"
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventLocation"
          description: Место успешно создано
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Создать новую локацию. Allowed roles=[ADMIN]
      tags:
      - Locations
  "/locations/{locationId}":
    delete:
      description: |-
        Удалить локацию по идентификатору.

        Изменения по итерациям:
        - Итерация 1: добавить endpoint для удаления локации по locationId. Если локация не найдена, должен возвращаться 404.
        - Итерация 2: endpoint теперь вызывается только с Bearer JWT; доступ разрешен только роли ADMIN.
        - Итерация 3: добавить ограничение; нельзя удалять локацию, если на ней уже есть мероприятия. В этом случае должен возвращаться 400.
        - Итерация 5: изменена внутренняя логика после удаления; добавлена инвалидация кешей локаций.
      parameters:
      - description: Уникальный идентификатор местоположения
        in: path
        name: locationId
        required: true
        schema:
          example: 42
          format: int64
          type: integer
      responses:
        '204':
          description: Место успешно удалено
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка удалить локацию, на которой уже есть мероприятия.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '404':
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Сущность не найдена.
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Удалить локацию. Allowed roles=[ADMIN]
      tags:
      - Locations
    get:
      description: |-
        Получить локацию по идентификатору.

        Изменения по итерациям:
        - Итерация 1: добавить endpoint для получения локации по locationId. Если локация не найдена, должен возвращаться 404.
        - Итерация 2: endpoint теперь вызывается только с Bearer JWT; доступ разрешен ролям ADMIN и USER.
        - Итерация 5: изменена внутренняя логика чтения; используется cache-aside для локации по id с Redis.
      parameters:
      - description: Уникальный идентификатор местоположения
        in: path
        name: locationId
        required: true
        schema:
          example: 42
          format: int64
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventLocation"
          description: Успешный ответ
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '404':
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Сущность не найдена.
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Получить локацию по ID. Allowed roles=[ADMIN, USER]
      tags:
      - Locations
    put:
      description: |-
        Обновить локацию по идентификатору.

        Изменения по итерациям:
        - Итерация 1: добавить endpoint для обновления локации с валидацией request body. Если локация не найдена, должен возвращаться 404.
        - Итерация 2: endpoint теперь вызывается только с Bearer JWT; доступ разрешен только роли ADMIN.
        - Итерация 3: добавить ограничение; нельзя уменьшать capacity, если это нарушает ограничения уже существующих мероприятий на этой локации. В этом случае должен возвращаться 400.
        - Итерация 5: изменена внутренняя логика после обновления; добавлена инвалидация кешей локаций.
      parameters:
      - description: Уникальный идентификатор местоположения
        in: path
        name: locationId
        required: true
        schema:
          example: 42
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            example:
              name: Новое имя локации
              address: Новый адрес
              capacity: 20000
              description: Новое описание
            schema:
              "$ref": "#/components/schemas/EventLocation"
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/EventLocation"
          description: Место успешно обновлено
        '400':
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Запрос с невалидными данными. Например, попытка занять занятый
            логин или некорректные параметры запроса.
        '401':
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Аутентификация отсутствует или не удалась
        '403':
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: У пользователя недостаточно прав для выполнения этой операции
        '404':
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Сущность не найдена.
        '500':
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: '2024-12-03T10:15:30'
              schema:
                "$ref": "#/components/schemas/ErrorMessageResponse"
          description: Внутренняя ошибка сервера
      summary: Обновить локацию. Allowed roles=[ADMIN]
      tags:
      - Locations
...
components:
...
  schemas:
    ErrorMessageResponse:
      example:
        message: Сообщение об ошибке
        detailedMessage: Более детальное сообщение об ошибке
        dateTime: '2024-12-03T10:15:30'
      properties:
        message:
          description: Человеко-читаемое сообщение об ошибке
          type: string
        detailedMessage:
          description: Более детальное сообщение об ошибке. Например, содержимое exception
          type: string
        dateTime:
          description: Дата и время возникновения ошибки. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
      required:
      - dateTime
      - detailedMessage
      - message
      type: object
    EventLocation:
      example:
        id: 42
        name: Ледовый дворец
        address: г. СПб, Пятилеток 1
        capacity: 12300
        description: Спортивно-концертный комплекс в Санкт-Петербурге
      properties:
        id:
          description: Уникальный идентификатор локации
          format: int64
          readOnly: true
          type: integer
        name:
          description: Имя локации
          type: string
        address:
          description: Адрес локации
          type: string
        capacity:
          description: Вместительность локации
          minimum: 5
          type: integer
        description:
          description: Дополнительное описание локации
          type: string
      required:
      - address
      - capacity
      - name
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http